+++ src/ase2sprkkr/outputs/readers/dos.py
@@ -1,6 +1,6 @@
 """ Density of states (DOS) reader and result. """
 
-from ..task_result import TaskResult, KkrProcess
+from ..task_result import TaskResult, KkrProcessRunner
 from .default import DefaultOutputReader
 from ...common.decorators import cached_property
 from ...output_files.output_files import OutputFile
@@ -25,7 +25,7 @@ class DosResult(TaskResult):
       return OutputFile.from_file(self.dos_filename, try_only='dos')
 
 
-class DosProcess(KkrProcess):
+class DosProcessRunner(KkrProcessRunner):
   """ ARPES task output reader currently do nothing, just have a special
   result, that allow easy acces to spc output file """